pyproject.toml
[tool.poetry]
name = "techno"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
jupyter = "^1.0.0"
pandas = "^1.3.5"
numpy = "^1.22.0"
scikit-learn = "^1.0.2"
matplotlib = "^3.5.1"
seaborn = "^0.11.2"
tqdm = "^4.62.3"
pandarallel = "^1.5.4"
japanize-matplotlib = "^1.1.3"
ipykernel = "^6.6.0"
requests = "^2.26.0"
timm = "^0.4.12"
pytorch-lightning = "^1.5.7"
black = "^21.12b0"
flake8 = "^4.0.1"
isort = "^5.10.1"
python-box = "^5.4.1"
ttach = "^0.0.3"
grad-cam = "^1.3.7"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"


